Combine readers
closes #2
