+++ src/Basic.CompilerLog.Util/CompilerCall.cs
@@ -13,16 +13,18 @@ public sealed class CompilerCall
     public string? TargetFramework { get; }
     public bool IsCSharp { get; }
     public string[] Arguments { get; }
+    internal int? Index { get; }
 
     public bool IsVisualBasic => !IsCSharp;
     public string ProjectFileName => Path.GetFileName(ProjectFilePath);
 
-    internal CompilerCall(string projectFilePath, CompilerCallKind kind, string? targetFramework, bool isCSharp, string[] arguments)
+    internal CompilerCall(string projectFilePath, CompilerCallKind kind, string? targetFramework, bool isCSharp, string[] arguments, int? index)
     {
         ProjectFilePath = projectFilePath;
         Kind = kind;
         TargetFramework = targetFramework;
         IsCSharp = isCSharp;
         Arguments = arguments;
+        Index = index;
     }
 }
